requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==0.15.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pyzbar==0.1.9
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
redis==5.0.1
pytest==7.4.3
Werkzeug==3.0.1
